# specs/002-user-personalization-module/contracts/personalization-api.openapi.yaml
openapi: 3.0.3
info:
  title: User Personalization API
  description: |
    API for managing user personalization preferences and generating personalized/translated content.
    Extends the existing FastAPI backend with new endpoints for preference management and content generation.
  version: 1.0.0
  contact:
    name: Physical AI & Humanoid Robotics Book - Backend Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (TBD)
security:
  - BetterAuthSession: []
...
paths:
  /api/auth/profile/preferences:
    get:
      tags:
        - Preferences
      summary: Get user personalization preferences
      description: Retrieve the authenticated user's personalization preferences
      operationId: getPersonalizationPreferences
      security:
        - BetterAuthSession: []
      responses:
        '200':
          description: Successfully retrieved preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizationPreferencesResponse'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Preferences not found (user hasn't submitted preferences yet)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Preferences
      summary: Create or update personalization preferences
      description: Submit or update the authenticated user's personalization preferences. Creates preferences if they don't exist, updates if they do.
      operationId: upsertPersonalizationPreferences
      security:
        - BetterAuthSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalizationPreferencesInput'
      responses:
        '200':
          description: Preferences successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizationPreferencesResponse'
        '201':
          description: Preferences successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizationPreferencesResponse'
        '400':
          description: Bad request - invalid preference data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BetterAuthSession:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Better Auth session token stored in httpOnly cookie
  schemas:
    PersonalizationPreferencesInput:
      type: object
      required:
        - experience_level
        - learning_topics
        - learning_goals
        - content_mode
      properties:
        experience_level:
          type: string
          enum: [beginner, intermediate, advanced]
          description: User's experience level in robotics/programming
          example: "intermediate"
        learning_topics:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          minItems: 1
          maxItems: 10
          description: Topics the user wants to focus on
          example: ["ROS 2", "Computer Vision", "Path Planning"]
        learning_goals:
          type: string
          maxLength: 500
          description: User's learning goals (free text)
          example: "I want to build autonomous navigation systems for mobile robots"
        content_mode:
          type: string
          enum: [full, personalized]
          description: Preferred content delivery mode
          example: "personalized"
        urdu_translation_enabled:
          type: boolean
          default: false
          description: Whether to enable Urdu translation
          example: false
    PersonalizationPreferencesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        preferences:
          type: object
          properties:
            experience_level:
              type: string
              enum: [beginner, intermediate, advanced]
              example: "intermediate"
            learning_topics:
              type: array
              items:
                type: string
              example: ["ROS 2", "Computer Vision", "Path Planning"]
            learning_goals:
              type: string
              example: "I want to build autonomous navigation systems for mobile robots"
            content_mode:
              type: string
              enum: [full, personalized]
              example: "personalized"
            urdu_translation_enabled:
              type: boolean
              example: false
            preferences_submitted_at:
              type: string
              format: date-time
              example: "2025-12-01T10:30:00Z"
            preferences_last_updated_at:
              type: string
              format: date-time
              example: "2025-12-01T10:30:00Z"
            preferences_version:
              type: integer
              minimum: 1
              example: 1
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Unauthorized: Please log in to access this resource"
        error_code:
          type: string
          description: Machine-readable error code
          example: "UNAUTHORIZED"
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        error_code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "experience_level"
              message:
                type: string
                example: "Must be one of: beginner, intermediate, advanced"